orders/src/routes/__test__/new.test.ts
import request from 'supertest'
import {app} from '../../app'
import {Order} from '../../models/order'
import {Ticket} from '../../models/ticktes'
import {natsWrapper} from '../../nats-wrapper'
import {OrderStatus} from '@hari-ticket/common'
import mongoose from 'mongoose'

it('should return bad request error if ticket is not exist', async () => {
    const ticketId = new mongoose.Types.ObjectId();
    await request(app)
        .post('/api/orders')
        .set('Cookie', global.signin())
        .send({
            ticketId
        }).expect(404)
})

it('returns an error if the ticket is already reserved', async () => {
  const ticket = Ticket.build({
    title: 'concert',
    price: 20,
  });
  await ticket.save();
  const order = Order.build({
    ticket,
    userId: 'laskdflkajsdf',
    status: OrderStatus.Created,
    expiresAt: new Date(),
  });
  await order.save();

  await request(app)
    .post('/api/orders')
    .set('Cookie', global.signin())
    .send({ ticketId: ticket.id })
    .expect(404);
});

it('reserves a ticket', async () => {
  const ticket = Ticket.build({
    title: 'concert',
    price: 20,
  });
  await ticket.save();

  const response=await request(app)
    .post('/api/orders')
    .set('Cookie', global.signin())
    .send({ ticketId: ticket.id })
    expect(response.status).toEqual(201);
});
